+++ examples/src/test/scala/tour/QuickTourAdmin.scala
@@ -83,11 +83,11 @@ object QuickTourAdmin {
     indexAndInsert.results()
 
     // Find using the text index
-    collection.count(text("textual content -irrelevant")).printResults("Text search matches: ")
+    collection.countDocuments(text("textual content -irrelevant")).printResults("Text search matches: ")
 
     // Find using the $language operator
     val textSearch: Bson = text("textual content -irrelevant", TextSearchOptions().language("english"))
-    collection.count(textSearch).printResults("Text search matches (english): ")
+    collection.countDocuments(textSearch).printResults("Text search matches (english): ")
 
     // Find the highest scoring match
     collection.find(textSearch)